Add basic orchestrator which is responsible for starting and restarting a gameloop.

See merge request dinf/projets/h20/eq02-jdisgames/diep.io!41

+++ server/priv/repo/migrations/20200219191646_create_scores.exs
@@ -3,7 +3,7 @@ defmodule Diep.Io.Repo.Migrations.CreateScores do
 
   def change do
     create table(:scores) do
-      add :game_id, :integer, null: false
+      add :game_id, :bigint, null: false
       add :score, :integer, null: false
       add :user_id, references(:users, on_delete: :nothing)
